+++ src/css/styles.css
@@ -172,10 +172,21 @@ label {
   max-width: 90vw;
   font-size: 30px;
   -webkit-text-stroke: 1px black;
+  font-weight: bold;
 }
 
 #players span {
   font-size: 15px;
   color: red;
   -webkit-text-stroke: 0px black;
+}
+
+#start-button {
+  display: none;
+  margin: 10px;
+  padding: 20px;
+  width: 30vw;
+  min-width: 300px;
+  height: 100px;
+  background-color: #12af12;
 }
